Move reused code to common.rs
The same implementation of the `Cloud` enum was found across commands.
This commit puts this code in a common place.
