+++ crates/core/tedge/src/cli/connect/command.rs
@@ -1,4 +1,9 @@
-use crate::{cli::connect::jwt_token::*, cli::connect::*, command::Command, ConfigError};
+use crate::{
+    cli::connect::jwt_token::*,
+    cli::{common::Cloud, connect::*},
+    command::Command,
+    ConfigError,
+};
 use rumqttc::QoS::AtLeastOnce;
 use rumqttc::{Event, Incoming, MqttOptions, Outgoing, Packet};
 use std::path::{Path, PathBuf};
@@ -30,30 +35,6 @@ pub struct ConnectCommand {
 pub enum DeviceStatus {
     AlreadyExists,
     Unknown,
-}
-
-#[derive(Debug)]
-pub enum Cloud {
-    Azure,
-    C8y,
-}
-
-impl Cloud {
-    fn dependent_mapper_service(&self) -> SystemService {
-        match self {
-            Cloud::Azure => SystemService::TEdgeMapperAz,
-            Cloud::C8y => SystemService::TEdgeMapperC8y,
-        }
-    }
-}
-
-impl Cloud {
-    fn as_str(&self) -> &'static str {
-        match self {
-            Self::Azure => "Azure",
-            Self::C8y => "Cumulocity",
-        }
-    }
 }
 
 impl Command for ConnectCommand {
@@ -155,10 +136,9 @@ impl Command for ConnectCommand {
             if which("tedge-mapper").is_err() {
                 println!("Warning: tedge-mapper is not installed.\n");
             } else {
-                self.service_manager.as_ref().start_and_enable_service(
-                    self.cloud.dependent_mapper_service(),
-                    std::io::stdout(),
-                );
+                self.service_manager
+                    .as_ref()
+                    .start_and_enable_service(self.cloud.mapper_service(), std::io::stdout());
             }
         }
