+++ crates/core/tedge/src/cli/disconnect/cli.rs
@@ -1,4 +1,4 @@
-use crate::cli::disconnect::disconnect_bridge::*;
+use crate::cli::{common::Cloud, disconnect::disconnect_bridge::*};
 use crate::command::*;
 use tedge_config::system_services::service_manager;
